app4/pyproject.toml
[project]
name = "hello-langfuse-otel"
version = "0.1.0"
description = "LangChain + OpenAI chat demo with Langfuse and OpenTelemetry (Traceloop)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.3.0,<0.4.0",
    "langchain-openai>=0.2.0,<0.3.0",
    "traceloop-sdk>=0.50.0,<0.51.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["app.py"]
